feat: split useNotificationsRenewal hook
Implement logic to show notifications renewal message in a separate hook
